backend/requirements.txt
fastapi 
uvicorn[standard] 
python-multipart 
tensorflow-cpu  # Replaces tflite-runtime
numpy 
opencv-python-headless 
Pillow 
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pytz